[Reason-CLI]
Summary: Switching to a new approach. I really like the old approach,
creating shell script wrappers that appear to be the binaries, but there
was one problem - the editor plugins search for `which refmt` and then
climb up ../ from that. By making wrapper scripts, we move the
"binaries" away from the other files around their install locations -
which is how we would find the editor support plugins.

Test Plan:

Reviewers:

CC: