Change directory to PACKAGE_ROOT before moving.
@andreypopp

I found the issue that was only happening on WSL but I don't fully understand what's happening. For some reason, a problem surfaces on WSL, where `mv` uses the *current* working directory's permissions - which is super unintuitive. So the fix seems to be to move to the correct package root directory before doing any move operations.